fix(editor) Fix circular dependency. - Moved `updateFileIfPossible` to a new file so that it doesn't cause   a circular dependency issue that only seems to arise with Vite.
